requirements.txt
-i https://pypi.org/simple
attrs==20.1.0
autobahn==20.7.1
automat==20.2.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
constantly==15.1.0
cryptography==3.0
dateparser==0.7.6
flask==1.1.2
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
investpy==1.0
itsdangerous==2.0.0a1
jinja2==3.0.0a1
lxml==4.5.2
markupsafe==2.0.0a1
multitasking==0.0.9
numpy==1.19.1
pandas==1.1.1
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.20
pyhamcrest==2.0.2
pyopenssl==19.1.0
python-binance==0.7.5
python-dateutil==2.8.1
pytz==2020.1
regex==2020.7.14
requests==2.24.0
scipy==1.5.2
selenium==4.0.0a6.post2
service-identity==18.1.0
six==1.15.0
twisted==20.3.0
txaio==20.4.1
tzlocal==2.1
unidecode==1.1.1
urllib3[secure]==1.25.10
werkzeug==1.0.1
yfinance==0.1.54
zope.interface==5.1.0
